Added frontend patches
