+++ src/frontend/containers/HomePage.tsx
@@ -1,8 +1,7 @@
 import React, { FC, useEffect, useState } from 'react';
 import { Path } from '../../shared/models/Path';
-import { NewPathButton } from '../components/NewPathButton';
-import { PathLink } from '../components/PathLink';
 import { captureError } from '../utils';
+import { PathCard } from '../components/PathCard';
 
 export const HomePage: FC = () => {
   const [loading, setLoading] = useState(true);
@@ -20,9 +19,10 @@ export const HomePage: FC = () => {
 
   return (
     <>
-      {!loading && paths.map((path) => <PathLink key={path.id} path={path} />)}
-      <h1>Hello, world! You are on / via React Router.</h1>
-      <NewPathButton />
+      <h1>Welcome to TrailBlazer</h1>
+      <p>Choose a Path below and follow along with each Waypoint.</p>
+      <h2>Public Paths:</h2>
+      {!loading && paths.map((path) => <PathCard key={path.id} path={path} />)}
     </>
   );
 };
